Fix logging gadget for replay speed setting.
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@2969 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ ground/openpilotgcs/src/plugins/logging/logfile.h
@@ -27,7 +27,7 @@ public:
     bool stopReplay();
 
 public slots:
-    void setReplaySpeed(int val) { playbackSpeed = pow(10,((double) val)/100); qDebug() << playbackSpeed; };
+    void setReplaySpeed(double val) { playbackSpeed = pow(10,(val)/100); qDebug() << playbackSpeed; };
     void pauseReplay();
     void resumeReplay();
