+++ ground/openpilotgcs/src/plugins/logging/logginggadgetwidget.cpp
@@ -58,7 +58,7 @@ void LoggingGadgetWidget::setPlugin(LoggingPlugin * p)
     connect(m_logging->playButton, SIGNAL(clicked()), scpPlugin, SLOT(startPlotting()));
     connect(m_logging->pauseButton,SIGNAL(clicked()),p->getLogfile(),SLOT(pauseReplay()));
     connect(m_logging->pauseButton, SIGNAL(clicked()), scpPlugin, SLOT(stopPlotting()));
-    connect(m_logging->playbackSpeed,SIGNAL(valueChanged(int)),p->getLogfile(),SLOT(setReplaySpeed(int)));
+    connect(m_logging->playbackSpeed,SIGNAL(valueChanged(double)),p->getLogfile(),SLOT(setReplaySpeed(double)));
     void pauseReplay();
     void resumeReplay();
 }
